{{cookiecutter.app_name}}/requirements.txt
git+http://github.com/mitsuhiko/flask.git
flask-classy
flask-sqlalchemy
flask-socketio
invoke